+++ source/ProcessManager.Orchestrations/Processes/BRS_023_027/V1/Activities/EnqueActorMessagesStep/EnqueueActorMessagesActivity_Brs_023_027_V1.cs
@@ -45,6 +45,6 @@ internal class EnqueueActorMessagesActivity_Brs_023_027_V1(
 
     public record ActivityInput(
         OrchestrationInstanceId InstanceId,
-        CalculatedDataForCalculationTypeV1 CalculatedData,
+        CalculationEnqueueActorMessagesV1 CalculatedData,
         Guid IdempotencyKey);
 }