+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,10 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
 
+## Version 0.17.4
+
+- Rename `NotifyEnqueueFinishedV1` to `CalculationEnqueueActorMessagesCompletedNotifyEventV1`.
+- Rename `CalculatedDataForCalculationTypeV1` to `CalculationEnqueueActorMessagesV1`.
+
 ## Version 0.17.3
 
 - Updated `MeteredDataForMeteringPointAcceptedV1`